feat: support passing an fga.mod file for the --file argument
This PR implements the following that allows us to begin accepting an fga.mod file in the CLI

* `ReadFromInputFileOrArg`/`ReadFromFile` now handles when an `fga.mod` is passed and switch the
format to `ModelFormatModular`
* `ReadModelFromString` now calls `ReadModelFromModFGA` when passed an input that is of type
`ModelFormatModular`
* The new function `ReadModelFromModFGA` will handle parsing the `fga.mod` file via language,
reading in the files specified in that file, then passing these off to language to transform into
the overall model.

In order to validate this I also added handling into the `fga model transform` command that will
print out the full DSL representation of the provided model.

Closes #266, #267, #268